fix(main/profile): harden CSP and record moderation actor
Follow-up to the review of #67.

Security:
- Drop https://unpkg.com from script-src. It was only there for the
  maplibre-gl CDN script on the (disabled) arcades map, and allowlisting an
  unpinned, SRI-less CDN alongside 'unsafe-inline' is a broad XSS and
  supply-chain surface. The Script tag in src/components/db/arcades.tsx is
  left in place; self-hosting maplibre-gl is a separate change.
- Stop hardcoding cdn.tomomai.lol / cdn.cn.tomomai.lol in img-src. All asset
  origins are derived from NEXT_PUBLIC_R2_URL, NEXT_PUBLIC_R2_URL_CN and
  RENDER_PUBLIC_URL again. headers() is evaluated at build time, so a missing
  value now logs a build warning instead of silently shrinking the policy.

Moderation:
- Add profile_reports.resolvedByUserId (nullable, ON DELETE SET NULL) so
  dismiss/removeDescription record which moderator acted, and surface the
  resolver in the admin dialog. adminProcedure authenticates with a shared
  bearer token, so this is null when no session accompanies the request.
- Match the pending-report unique violation through wrapped driver errors.
  Drizzle nests the pg error under `cause`, so duplicate reports were
  surfacing as 500s instead of CONFLICT.
- Check the reporter's rate limit before the target lookup so probing for
  user ids costs quota.

Profile UI:
- Extract ProfileAboutSection and the visibility dialog out of info-card.tsx
  (904 -> 501 lines) into profile-about-section.tsx and
  profile-visibility-dialog.tsx.
- Lazy-load MarkdownEditor and ProfileReportDialog. The editor is owner-only
  and the report dialog visitor-only, but both shipped in the chunk every
  profile visitor downloaded. @tomomai/markdown now exposes an ./editor
  subpath so the dynamic import targets just that module.

Smaller fixes:
- Localize the video embed's "Load video" button and region label via
  createVideoEmbedExtension(labels); they were hardcoded English.
- measureMarkdown uses TextEncoder unconditionally (Buffer is absent on edge).
- Keep the internal data-standalone-url marker off the rendered <p>.
- updateProfileDescription takes the refined schema as its tRPC input instead
  of re-validating in the resolver.

The 0016 migration was regenerated (unapplied) and must still be applied
before deploy: profileDescription is read on every public profile load.
