+++ apps/main/drizzle-pg/0016_blue_pretty_boy.sql
@@ -10,12 +10,14 @@ CREATE TABLE "profile_reports" (
 	"status" "profile_report_status" DEFAULT 'pending' NOT NULL,
 	"createdAt" timestamp (0) DEFAULT now() NOT NULL,
 	"resolvedAt" timestamp (0),
+	"resolvedByUserId" text,
 	"resolutionNote" varchar(1000)
 );
 --> statement-breakpoint
 ALTER TABLE "user" ADD COLUMN "profileDescription" text;--> statement-breakpoint
 ALTER TABLE "profile_reports" ADD CONSTRAINT "profile_reports_reporterUserId_user_id_fk" FOREIGN KEY ("reporterUserId") REFERENCES "public"."user"("id") ON DELETE cascade ON UPDATE no action;--> statement-breakpoint
 ALTER TABLE "profile_reports" ADD CONSTRAINT "profile_reports_targetUserId_user_id_fk" FOREIGN KEY ("targetUserId") REFERENCES "public"."user"("id") ON DELETE cascade ON UPDATE no action;--> statement-breakpoint
+ALTER TABLE "profile_reports" ADD CONSTRAINT "profile_reports_resolvedByUserId_user_id_fk" FOREIGN KEY ("resolvedByUserId") REFERENCES "public"."user"("id") ON DELETE set null ON UPDATE no action;--> statement-breakpoint
 CREATE INDEX "profile_reports_status_createdat_idx" ON "profile_reports" USING btree ("status","createdAt");--> statement-breakpoint
 CREATE INDEX "profile_reports_targetuserid_idx" ON "profile_reports" USING btree ("targetUserId");--> statement-breakpoint
 CREATE UNIQUE INDEX "profile_reports_pending_reporter_target_idx" ON "profile_reports" USING btree ("reporterUserId","targetUserId") WHERE "profile_reports"."status" = 'pending';
