+++ packages/markdown/package.json
@@ -7,6 +7,10 @@
     ".": {
       "import": "./src/index.ts",
       "types": "./src/index.ts"
+    },
+    "./editor": {
+      "import": "./src/editor.tsx",
+      "types": "./src/editor.tsx"
     }
   },
   "main": "./src/index.ts",
