+++ packages/markdown/src/index.ts
@@ -1,7 +1,8 @@
 export { MarkdownEditor } from "./editor";
 export type { MarkdownEditorLabels, MarkdownEditorProps } from "./editor";
 export { MarkdownContent, PROFILE_MARKDOWN_POLICY } from "./content";
-export { videoEmbedExtension, parseSupportedVideoUrl } from "./video";
+export { videoEmbedExtension, createVideoEmbedExtension, parseSupportedVideoUrl } from "./video";
+export type { VideoEmbedLabels } from "./video-embed";
 export { markdownBaseComponents, ExternalMarkdownLink } from "./trusted";
 export { measureMarkdown, validateMarkdownSize } from "./size";
 export type { MarkdownExtension, MarkdownLimits, MarkdownPolicy, MarkdownSize } from "./types";
